New ISprite format extended by all graphics
Palettes have been somewhat implemented but are not put to use yet.
PaletteDictionary definitions are not final. `palette_shader.fx` is
unused at the moment and is not up to date compared to
`palette_lerp_shader.fx`.

ISprite drawing code is a little messy and will be cleaned up before
this is merged with wpf-develop.

Animations are stored in Resources<ISprite> AND Resources<Animation> to
maintain legacy animation support. I may change it so all sprite
variants have a subcategory as well.

ColorGroup sprites (For entitiy variants) have not been implemented yet.