Working Conscripts and Tile Progress
Conscripts are now fully fixed up and reimplemented. Prefixes have been
replaced with a more robust system to specify prefixes by linking
command names together.

Added Conscript "types" that behave that are composed of parameter
overloads just like commands. Types are only present in the
CommandReferenceParam and are stored normally in the CommandParam when
parsing.

Lots of *very slow* progress on building up tilesets to eventually
combine with the palette system.