Progress on Tile Implementation
Started using the Conscript Designer with the actual progress.

Numerous designer and conscript bugs have been fixed.

Added BASIC CLONE, COLORSTYLE, and COLORMULTISTYLE commands.

Changed when editor is initialized so that it's run on a dispatcher (UI)
thread.

Editor now asks for the new name of resources that are missing. You can
filter out resources in a list, choose to remove the resource instead,
or remove all remaining resources.