Fix Tile Cover Surface Issue
* Added more NotSurface flags to certain tiles. Most notably the
PullHandle which would invalidate the solid wall behind it.
* Added virtual bool IsStatic property to Tile class. By default this
returns false if any flag that allows tile movement/desctruction (except
Digable) exists. It also returns false if the tile is following a path.
* IsSurface now also requires that IsStatic is true.
* A few tiles now override IsStatic to force it to false. IsStatic has
no other uses, but in the future it might be useful to know if the tile
will ever leave it's spot.
* Add Tile.surfaceStatic, if set to false, the tile will not need
IsStatic to be true in order to be considered a surface. Implemented
soley for TileMonsterFlyingTile at the moment as it's not Static but has
full surface capabilities until then. (In OOS L-7, flying tiles lay on
top of holes) Again, IsStatic is only overriden to false for future use.