Areas, Monster Respawning, & Parenting
Areas:
* Dungeons have been replaced with Areas.
* Levels are now required to have an area. If the area is not valid, the
returned area will be a default area constructed by the world.
* GameControl and RoomControl now contain an AreaControl that manages
the current area.

Monster (Re)spawing
* AreaControl now contains RespawnManager that keeps track of all
monsters that are not permenantly dead.
* Areas contain monster spawn and respawn settings. Respawn settings can
be overridden per room.
* Monster's "dead" property is now only assigned if the monster is
permentantly dead.
* Red Zol and Gels now behave properly when it comes to room clearing
and respawn clearing.
* Monsters now have unique MonsterID property that can be assigned to
them in order to share death. Example: Used with Red Zol and Gels.

Parenting:
* Levels and rooms can now assign a parent level. Rooms can also change
the location of the parent room in the parent level.
* Children of parented rooms now contain any "shared" tiles within the
parent room. These tiles are the actual objects in the parent room and
thus share properties and settings.
* Added SHARED bool shared command to TileDataSR.
* Monsters and Rewards/Chests are shared by default while all other
tiles will need the property to be enabled.
* Shared tiles can optionally be viewed in the World Editor.
* Shared rooms will also share monster respawning. This is needed in
instances like Seasons where each season is a different level but has
the same monsters.

Misc:
* A few non-tile undo actions have been fixed so the created object is
now stored in the action. This prevents issues with property objects
later on.
* Added TextReader alignment. <align-left>, <align-center>,
<align-right>.
* Added area name display optionality for warp actions. (Specify for the
source warp, not the destination).
* Most room commands for accessing tiles and actions now have a boolean
for checking parented tiles too.
* Room.ActionData property has been phased out in favor of
GetActionTiles(), GetActionTileAt(), and ActionCount.
* DOCUMENTS TileDataSR command renamed to DOCUMENT
* Added LOCKPROP command to hide and make a property readonly in the
editor.
* AreaControl is now accessible as an interface in the ScriptObject
class.
* Added MapScreen base class for OnOpen event.