BaseResourceData + Item/Ammo/RewardData
* BaseResourceData is now the base type for all classes stored in
resources that are to be used to create a type at runtime. BaseTileData
now extends this too.
* ResourceDataInitializing and ResourceDataDrawing are now used to store
delegates to initialization and drawer functions for resource data
types.
* GameData.LoadInventory/Resources is no longer called during game start
and is now always called within GameData.Initialize.
Inventory/RewardManager.Initialize is now what should be called during
game start to load the data into Items/Ammo/Rewards.
* BaseTileData.Name (the resource name) has been renamed to ResourceName
to prevent conflicts with AmmoData.Name (the display name of the ammo).
* Rewards and Invenotry are still loaded when in the editor or designer
because they don't really do anything themselves or require GameControl.
This is also important because Rewards created from Items do not have
RewardData associated with them, so these rewards would not be able to
be looked up from within Resources.
* Removed all commented out Legacy Item/Ammo/Reward code.
* Fixed PalettedSpriteDatabase preloading always failing.
* Fixed Image.SaveAsPngAndSize writing the file size in the wrong place.