Many New Utilities and Extensions
* LevelEditorExtensions: Tons of new extensiosn for working with level
tile coords, level pixel positions, etc. Also phases out LevelTileCoord
struct and all uses for it.
* BinaryCounter: A class to keep track of a count at a starting position
then save the count to the starting postion in the stream. See
references to class for more info.
* Tiles, Rooms, and Levels now have many more helpful dimension and
position properties.
* PLEASE NOTE: BaseTileData(Instance).Size has been renamed to TileSize.
There is no Size function now so it shouldn't cause any overlooked code.
* GMath now has FloorDiv, and CeilingDiv to divide by integers and then
roud the value instead of just truncating it by default.
* GMath now has Sign2 which will return 1 if the Sign's return would be
0.
* Added Rounding And Sign functions with 'I' postfix. These functions
always return either an int or Point2I to space and avoiding the need to
manually cast.
* GMath class split into multiple classes.
* ZeldaEditorLegacy removed from solution but not yet deleted.
* Gmath.PlusDir variants now renamed to WrapAngle.
* Added Rectangle2I.FromEndPointsOne which will add (1, 1) to the size
of the rectangle.